% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_eggNOG.R
\name{load.eggNOG}
\alias{load.eggNOG}
\title{Load and Process eggNOG Output}
\usage{
load.eggNOG(file, col_names = T, delim = "\\t", comment = "##", ...)
}
\arguments{
\item{file}{The path to the eggNOG output file.}

\item{col_names}{Logical indicating whether the file has column names. Defaults to TRUE.}

\item{delim}{The delimiter used in the file. Defaults to tab ('\t').}

\item{comment}{The character used to indicate comments in the file. Defaults to '##'.}

\item{...}{Additional arguments passed to readr::read_delim.}
}
\value{
A data frame with the first column renamed to 'Gene'.
}
\description{
This function loads an eggNOG output file and processes it by renaming the first column to 'Gene'.
It uses the readr package to read the delimited file and dplyr to rename the columns.
}
\examples{
\dontrun{
# Load an eggNOG output file
eggno <- load.eggNOG("path/to/eggNOG_output.txt")

# View the first few rows of the data frame
head(eggno)
}
}
