% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_CAZy.R
\name{parse.eggNOG.CAZy}
\alias{parse.eggNOG.CAZy}
\title{Parse eggNOG CAZy Annotations}
\usage{
parse.eggNOG.CAZy(eggNOG.file, ...)
}
\arguments{
\item{eggNOG.file}{The path to the eggNOG output file.}

\item{...}{Additional arguments passed to load.eggNOG.}
}
\value{
A list containing two data frames:
\item{TERM2GENE}{A data frame with columns 'CAZy' and 'Gene', where each row represents a unique CAZy term associated with a gene.}
\item{TERM2NAME}{A data frame with columns 'CAZy' and 'name', where each row represents a unique CAZy term and its corresponding name.}
}
\description{
This function takes an eggNOG output file and parse CAZy annotations,
filtering out genes without CAZy annotations, and returning a list of two data frames:
one mapping CAZy terms to genes and another mapping CAZy terms to their names.
}
\examples{
\dontrun{
# Parse an eggNOG output file for CAZy annotations
CAZy.res <- parse.eggNOG.CAZy("path/to/eggNOG_output.txt")

# View the TERM2GENE data frame
print(CAZy.res$TERM2GENE)

# View the TERM2NAME data frame
print(CAZy.res$TERM2NAME)
}
}
