% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_KEGGEC.R
\name{parse.eggNOG.EC}
\alias{parse.eggNOG.EC}
\title{Parse eggNOG EC Number Annotations}
\usage{
parse.eggNOG.EC(eggNOG.file, database = "KEGG")
}
\arguments{
\item{eggNOG.file}{The path to the eggNOG output file.}

\item{database}{A character string specifying the database to use for EC number lookup.
Defaults to "KEGG". Can be set to "GotEnzymes" for slower, real-time lookup.}
}
\value{
A list of two data frames:
\item{TERM2GENE}{A data frame mapping EC numbers to genes.}
\item{TERM2NAME}{A data frame mapping EC numbers to their names.}
}
\description{
This function processes an eggNOG output file to parse EC number annotations.
It filters out genes without EC numbers, maps EC numbers to their corresponding genes,
and returns a list of data frames that map EC numbers to genes and their names.
The function can use either the KEGG database or the GotEnzymes database for EC number annoate.
}
\examples{
\dontrun{
# Parse an eggNOG output file for EC number annotations using the KEGG database
ec_annotations <- parse.eggNOG.EC("path/to/eggNOG_output.txt")

# Parse an eggNOG output file for EC number annotations using the GotEnzymes database (very slowly)
ec_annotations_gote <- parse.eggNOG.EC("path/to/eggNOG_output.txt", database = "GotEnzymes")

# View the TERM2GENE data frame
print(ec_annotations$TERM2GENE)

# View the TERM2NAME data frame
print(ec_annotations$TERM2NAME)
}
}
