% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_GO.R
\name{parse.eggNOG.GO}
\alias{parse.eggNOG.GO}
\title{Parse eggNOG GO Annotations}
\usage{
parse.eggNOG.GO(eggNOG.file, ...)
}
\arguments{
\item{eggNOG.file}{The path to the eggNOG output file.}

\item{...}{Additional arguments passed to load.eggNOG.}
}
\value{
A list of data frames:
\item{TERM2GENE.CC}{A data frame mapping GO terms (Cellular Component) to genes.}
\item{TERM2GENE.BP}{A data frame mapping GO terms (Biological Process) to genes.}
\item{TERM2GENE.MF}{A data frame mapping GO terms (Molecular Function) to genes.}
\item{TERM2GENE.ALL}{A data frame mapping all GO terms to genes.}
\item{TERM2NAME}{A data frame mapping GO terms to their names.}
}
\description{
This function processes an eggNOG output file to parse Gene Ontology (GO) annotations.
It filters out genes without GO annotations, maps GO terms to their corresponding genes and descriptions,
and returns a list of data frames that categorize genes by GO level (Cellular Component, Biological Process, Molecular Function).
}
\examples{
\dontrun{
# Parse an eggNOG output file for GO annotations
go_annotations <- parse.eggNOG.GO("path/to/eggNOG_output.txt")

# View the TERM2GENE.CC data frame
print(go_annotations$TERM2GENE.CC)

# View the TERM2NAME data frame
print(go_annotations$TERM2NAME)
}
}
