% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_KEGGP.R
\name{parse.eggNOG.KEGGP}
\alias{parse.eggNOG.KEGGP}
\title{Parse eggNOG KEGG Pathway Annotations}
\usage{
parse.eggNOG.KEGGP(eggNOG.file, species = NULL, ...)
}
\arguments{
\item{eggNOG.file}{The path to the eggNOG output file.}

\item{species}{A character string specifying the species for pathway filtering.
Defaults to NULL, which means no species-specific filtering is applied.supported organism listed in 'https://www.genome.jp/kegg/catalog/org_list.html'}

\item{...}{Additional arguments passed to load.eggNOG.}
}
\value{
A list of two data frames:
\item{TERM2GENE}{A data frame mapping KEGG pathway identifiers to genes.}
\item{TERM2NAME}{A data frame mapping KEGG pathway identifiers to their names.}
}
\description{
This function processes an eggNOG output file to parse KEGG pathway annotations.
It filters out genes without KEGG pathway annotations, maps KEGG pathway identifiers to their corresponding genes,
and returns a list of data frames that map KEGG pathways to genes and their names.
Optionally, it can filter pathways by a specified species.
}
\examples{
\dontrun{
# Parse an eggNOG output file for KEGG pathway annotations without species filtering
keggp_annotations <- parse.eggNOG.KEGGP("path/to/eggNOG_output.txt")

# Parse an eggNOG output file for KEGG pathway annotations with species filtering
keggp_annotations_species <- parse.eggNOG.KEGGP("path/to/eggNOG_output.txt", species = "hsa")

# View the TERM2GENE data frame
print(keggp_annotations$TERM2GENE)

# View the TERM2NAME data frame
print(keggp_annotations$TERM2NAME)
}
}
