% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_KEGGR.R
\name{parse.eggNOG.KEGGR}
\alias{parse.eggNOG.KEGGR}
\title{Parse eggNOG KEGG Reaction Annotations}
\usage{
parse.eggNOG.KEGGR(eggNOG.file, species = NULL, ...)
}
\arguments{
\item{eggNOG.file}{The path to the eggNOG output file.}

\item{species}{A character string specifying the species for reaction filtering.
Defaults to NULL, which means no species-specific filtering is applied.Supported organism listed in 'https://www.genome.jp/kegg/catalog/org_list.html'}

\item{...}{Additional arguments passed to load.eggNOG.}
}
\value{
A list of two data frames:
\item{TERM2GENE}{A data frame mapping KEGG reaction identifiers to genes.}
\item{TERM2NAME}{A data frame mapping KEGG reaction identifiers to their names.}
}
\description{
This function processes an eggNOG output file to parse KEGG reaction annotations.
It filters out genes without KEGG reaction annotations, maps KEGG reaction identifiers to their corresponding genes,
and returns a list of data frames that map KEGG reactions to genes and their names.
Optionally, it can filter reactions by a specified species.
}
\examples{
\dontrun{
# Parse an eggNOG output file for KEGG reaction annotations without species filtering
keggr_annotations <- parse.eggNOG.KEGGR("path/to/eggNOG_output.txt")

# Parse an eggNOG output file for KEGG reaction annotations with species filtering
keggr_annotations_species <- parse.eggNOG.KEGGR("path/to/eggNOG_output.txt", species = "hsa")

# View the TERM2GENE data frame
print(keggr_annotations$TERM2GENE)

# View the TERM2NAME data frame
print(keggr_annotations$TERM2NAME)
}
}
